+++ tester/include/malloc_fail.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   malloc_fail.hpp                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: neddykun <neddykun@student.42.fr>          +#+  +:+       +#+        */
+/*   By: libft-tester <opensource@example.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2026/07/06 22:54:20 by neddykun          #+#    #+#             */
-/*   Updated: 2026/07/06 22:54:21 by neddykun         ###   ########.fr       */
+/*   Created: 2026/07/06 22:54:20 by libft-tester          #+#    #+#             */
+/*   Updated: 2026/07/06 22:54:21 by libft-tester         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */